Split tests to separate files and add new ones in test_forms
